+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This file contains the list of changes made to pyjoulescope_ui.
   * Adapted minimum x width based upon signal frequencies #284
   * Fixed annotation load sometimes not working due to race condition #283
   * Fixed duplicate annotation load for JLS files when toggling views
+  * Added export visible data to CSV  #274
 
 
 ## 1.1.10